// src/generated/types/StudentDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type StudentDto = {
  /**
   * @type string
   */
  id: string
  /**
   * @type string
   */
  name: string
  /**
   * @type string
   */
  username: string
  /**
   * @type number
   */
  seatNo: number
  /**
   * @type number
   */
  level: number
  /**
   * @type string
   */
  program: string
  /**
   * @description Student photo ID
   * @type string
   */
  photo: string
}